Reference

In this reference, you will find a detailed overview of the package API.

Reference guides are technical descriptions of the machinery and how to operate it. Reference material is information-oriented.

— Diátaxis

In other words, you come here because you want to take a very close look at the code 🧐.

Content

Reference

Content

Exported functions

Internal functions

Exported functions

EnergySamplers.ConditionalSampler — Method

ConditionalSampler(
    X::AbstractArray, y::AbstractArray;
    batch_size::Int,
    max_len::Int=10000, prob_buffer::AbstractFloat=0.95
)

Outer constructor for ConditionalSampler.

JointEnergyModels.JointEnergyClassifier — Type

The JointEnergyClassifier struct is a wrapper for a JointEnergyModel that can be used with MLJFlux.jl.

JointEnergyModels.class_loss — Method

class_loss(jem::JointEnergyModel, x, y)

Computes the classification loss.

JointEnergyModels.gen_loss — Method

gen_loss(jem::JointEnergyModel, x)

Computes the generative loss.

JointEnergyModels.generate_conditional_samples — Method

generate_conditional_samples(model, rule::Flux.Optimise.AbstractOptimiser, n::Int, y::Int; kwargs...)

A convenience function for generating conditional samples for a given model, sampler and sampling rule. If n is missing, then the sampler's batch_size is used. The conditioning value y needs to be specified.

JointEnergyModels.generate_samples — Method

generate_samples(jem::JointEnergyModel, n::Int; kwargs...)

A convenience function for generating samples for a given energy model. If n is missing, then the sampler's batch_size is used.

JointEnergyModels.loss — Method

loss(jem::JointEnergyModel, x, y; agg=mean)

Computes the total loss.

Internal functions

JointEnergyModels.get_samples — Method

get_samples(jem::JointEnergyModel, x)

Gets samples from the sampler buffer.

JointEnergyModels.reg_loss — Method

reg_loss(jem::JointEnergyModel, x)

Computes the regularization loss.